+++ packages/gcn/src/instructions/formats/mimg.rs
@@ -60,7 +60,7 @@ impl FromBits<4> for DMask {
 impl<R: Reader> ParseInstruction<R> for MIMGInstruction {
     fn parse(token: u32, reader: R) -> Result<Self, anyhow::Error> {
         let token = combine(token, reader)?;
-        Ok(MIMGInstruction::from_bits(token as usize))
+        Ok(MIMGInstruction::from_bits(token))
     }
 }
